Finished Scene 6
As I complete this scene, I realize that I cannot complete all the scenes as I wanted (before the deadline, at least). Therefore, I will trim the number of scenes from 10 to 7. I believe that quality is better than quantity for the scenes.

For Scene 6, I brought in a font file and created the Choice_circle class.