Finished Revision Process, Finished Project!
While I did try to see how I could improve the firework and raindrops, I realized that I would have to change their foundations. That is, the classes for firework and raindrop.

As for animating certain things, I decided to bypass that since I decided to move away from p5.play. The sprite sheets required would take up too much time.

I was able to improve Scene 5, with the impending eclipse. By working on the fills and position of the moon, I made the scene lighting more like that of an eclipse.